requirements.txt
astral==3.2
python-dotenv==1.0.0
prometheus-client==0.19.0
holidays==0.38
loguru==0.7.2
aiohttp==3.9.1
pytz==2024.1
pyyaml==6.0.1